# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor Management & Appointment Booking API",
        "description": "REST API for the Doctor Management and Simple Appointment Booking System.\n\nRoles: Management (admin), Doctor, Regular user (patient).\n\nAuthentication uses Laravel Sanctum bearer tokens. Obtain a token from POST /api/v1/login, then send it as `Authorization: Bearer <token>`.",
        "contact": {
            "name": "Support",
            "email": "mebrahimi405@yahoo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development"
        }
    ],
    "paths": {
        "/api/v1/admin/doctors": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all doctors",
                "description": "Display a listing of the resource.",
                "operationId": "adminListDoctors",
                "responses": {
                    "200": {
                        "description": "List of doctors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DoctorManagementItem"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/doctors/{doctor}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve / reject / deactivate a doctor (PUT)",
                "operationId": "adminUpdateDoctorStatusPut",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated doctor profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve / reject / deactivate a doctor",
                "description": "Change a doctor's status. Use 'active' to approve, 'suspended' to deactivate.",
                "operationId": "adminUpdateDoctorStatus",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated doctor profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/doctors/profile": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all doctor profiles (paginated)",
                "description": "Display a listing of the resource.",
                "operationId": "adminListDoctorProfiles",
                "responses": {
                    "200": {
                        "description": "Paginated list of doctor profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedDoctorProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/doctors/profile/{profile}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit a doctor profile (PUT)",
                "operationId": "adminUpdateDoctorProfilePut",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorProfileAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated doctor profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit a doctor profile",
                "description": "Update the specified resource in storage.",
                "operationId": "adminUpdateDoctorProfile",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorProfileAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated doctor profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfile"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/doctor/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new doctor",
                "description": "Register a doctor account. The account is created with a 'pending' status and cannot access the panel until approved by an admin.",
                "operationId": "registerDoctor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DoctorRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor registered (pending approval)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DoctorStatus": {
                "description": "Doctor profile status",
                "type": "string",
                "enum": [
                    "pending",
                    "active",
                    "rejected",
                    "suspended",
                    "vacation",
                    "archived"
                ]
            },
...
            "DoctorRegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Dr. John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "doctor@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    },
                    "specialty_id": {
                        "description": "Optional specialty ID",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateDoctorStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/DoctorStatus"
                    }
                },
                "type": "object"
            },
...
            "UpdateDoctorProfileAdminRequest": {
                "properties": {
                    "specialty_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "status": {
                        "$ref": "#/components/schemas/DoctorStatus"
                    },
                    "image": {
                        "type": "string",
                        "example": "doctors/avatar.jpg"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Updated biography."
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@test.com"
                    }
                },
                "type": "object"
            },
...
            "DoctorProfile": {
                "properties": {
                    "profile_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "specialty_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/DoctorStatus"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "doctors/avatar.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced cardiologist.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DoctorManagementItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "doctor@example.com"
                    },
                    "profile": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DoctorProfile"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginatedDoctorProfile": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DoctorProfile"
                        }
                    },
                    "links": {
                        "type": "object",
                        "example": {
                            "first": null,
                            "last": null,
                            "prev": null,
                            "next": null
                        }
                    },
                    "meta": {
                        "type": "object",
                        "example": {
                            "current_page": 1,
                            "per_page": 15,
                            "total": 15
                        }
                    }
                },
                "type": "object"
            },
...
            "RegisterResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Doctor registered successfully. Account is pending admin approval."
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "doctor_profile": {
                        "$ref": "#/components/schemas/DoctorProfileFull"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Unauthenticated - missing or invalid token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - the authenticated user does not have the required role/permission.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized action."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "No query results for model."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "email": [
                                            "The email field is required."
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Enter the token returned by /api/v1/login in the format \"Bearer <token>\".",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}